Update README.md
This file was good, but it wasn't good enough.  I changed this file in order to insure that it would aspire to greatness.  All other text files before and after this commit will pale in comparison to the majesty that is written herein.  This is the sum, the all, the perfect circle.  To gaze upon the contents of this .txt is to gaze into the reflection of one's own inadequacies, to weep for the utter lack of potential that you will soon realize rests within your own heart.  But do not fret, this .txt exists not to condemn the frailties of man, but to show us the heights of greatness that can be achieved.